trash_classifier_web/requirements.txt
Streamlit
Flask
tensorflow-cpu
numpy
pillow
matplotlib
scikit-learn
requests